Wrapper function, plotting the heatmap

This function is package internal and should not be executed directly by users.

heatmapPlotWrapper(MAT, axhline = NULL, titles = rep("", length(MAT)),
  bins = 1:(ncol(MAT[[1]])/length(MAT)), cex.lab = 12, cex.axis = 12,
  cex.legend = 12, xlab = "", ylab = "", Leg = TRUE, autoscale = TRUE,
  zmin = 0, zmax = 10, xlim = NULL, ln.v = TRUE, e = NULL, s = 0.01,
  indi = TRUE, o_min = NA, o_max = NA, colvec = NULL,
  colorspace = NULL, pointsize = 12, embed = FALSE, raster = FALSE,
  ylim = c(nrow(MAT[[1]]), 1), dendro = NULL, ...)

Arguments

MAT
- list of matrixes holding heatmap data

axhline
- locations of horizontal lines separating the clusters

titles
the sub-titles of heatmaps

bins
the x-axis indicates in heatmap

cex.lab
Axis labels font size in points, Defaults to 12

cex.axis
Axis numbers font size in points, defaults to 12

cex.legend
Keys labels font size in points, defaults to 12

xlab
label below x-axis

ylab
label below y-axis

autoscale
if TRUE the color keys will be auto scaled

zmin
global minimum value on color key, ignored if autoscale is TRUE

zmax
global maximum value on color key, ignored if autoscale is TRUE

xlim
the x limits (x1, x2) of the plot. Note that x1 > x2 is allowed and leads to a "reversed axis". The default value, NULL, indicates that the whole range present in plotset will be plotted.

ln.v
Determines if vertical guide line(s) should be plotted (TRUE) or ommitted (FALSE). For anchored plots 2 lines indicating the start and end of anchored distance are plotted.

e
Determines the end of anchored distance

s
The saturation value used to auto scale color key limits, defaults to 0.01

indi
If TRUE (defaults) the independent color keys will be plotted below heatmaps, if FALSE the common color key is shown rightmost

o_min
vector of length equal to number of sub heatmaps determining minimum value on color key for each sub plot, if NULL (default) or NA the global settings are used, ignored in indi is FALSE

o_max
vector of length equal to number of sub heatmaps determining maximum value on color key for each sub plot, if NULL (default) or NA the global settings are used, ignored in indi is FALSE

colvec
The vector of colors used to plot the lines and error estimate fields. If set value NULL (default) the automatically generated color values will be used. Accepted values are: vector of any of the three kinds of R color specifications, i.e., either a color name (as listed by colors()), a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a positive integer i meaning palette()[i]. See col2rgb.

colorspace
The colorspace of the heatmap, see grDevices

pointsize
The default font point size to be used for plots. Defaults to 12 (1/72 inch).

embed
Configures heatmap to be used embedded plot. Defaults to FALSE

raster
Uses raster graphics for heatmaps. Defaults to FALSE

ylim
Y axis limits. Defaults to c(nrow(MAT[[1]]),1)

dendro
Dendrogram object, will be plotted left to heatmaps. Defaults to NULL

leg
if TRUE plot the color key

Value

NULL