Modified image.plot function from "fields" package, the grey rectangle is used as background for heatmap, making NAs distinguishable from the data

#added following line of code: rect(usr[1], usr[3], usr[2], usr[4], col="grey") This function is package internal and should not be executed directly by users.

imPlot2(..., add = FALSE, nlevel = 64, horizontal = FALSE,
  legend.shrink = 0.9, legend.width = 1.2, legend.mar = ifelse(horizontal,
  3.1, 5.1), legend.lab = NULL, legend.line = 2, graphics.reset = FALSE,
  bigplot = NULL, smallplot = NULL, legend.only = FALSE,
  col = tim.colors(nlevel), lab.breaks = NULL, axis.args = NULL,
  legend.args = NULL, midpoint = FALSE, border = NA, lwd = 1,
  xinds = NULL, e = NULL, ylast = 0, afs = 1)

Value

NULL

Details

For more help see documentation of image.plot function: image.plot