SeqPlots - An interactive tool for visualizing NGS signals and sequence motif densities along genomic features using average plots and heatmaps.

SeqPlots is a tool for plotting next generation sequencing (NGS) based experiments' signal tracks, e.g. reads coverage from ChIP-seq, RNA-seq and DNA accessibility assays like DNase-seq and MNase-seq, over user specified genomic features, e.g. promoters, gene bodies, etc. It can also calculate sequence motif density profiles from reference genome. The data are visualized as average signal profile plot, with error estimates (standard error and 95 of heatmaps that can be sorted and clustered using hierarchical clustering, k-means algorithm and self organising maps. Plots can be prepared using R programming language or web browser based graphical user interface (GUI) implemented using Shiny framework. The dual-purpose implementation allows running the software locally on desktop or deploying it on server. SeqPlots is useful for both for exploratory data analyses and preparing replicable, publication quality plots. Other features of the software include collaboration and data sharing capabilities, as well as ability to store pre-calculated result matrixes, that combine many sequencing experiments and in-silico generated tracks with multiple different features. These binaries can be further used to generate new combination plots on fly, run automated batch operations or share with colleagues, who can adjust their plotting parameters without loading actual tracks and recalculating numeric values. SeqPlots relays on Bioconductor packages, mainly on rtracklayer for data input and BSgenome packages for reference genome sequence and annotations.

Details

Useful links:

Project web page: https://github.com/Przemol/seqplots

Online documentation: http://przemol.github.io/seqplots/

Project wiki: https://github.com/przemol/seqplots/wiki

Issue tracker: http://github.com/przemol/seqplots/issues

Bug reports: http://github.com/przemol/seqplots/issues/new

References

R project and Bioconductor:

R Core Team (2014). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL http://www.R-project.org/.

Bioconductor: Open software development for computational biology and bioinformatics R. Gentleman, V. J. Carey, D. M. Bates, B.Bolstad, M.Dettling, S. Dudoit, B. Ellis, L. Gautier, Y. Ge, and others 2004, Genome Biology, Vol. 5, R80. URL http://www.bioconductor.org/.

RStudio and Inc. (2014). shiny: Web Application Framework for R. R package version 0.10.1. http://shiny.rstudio.com/.

Other CRAN packages: digest, DBI,RSQLite, jsonlite, plotrix, fields, grid, kohonen, Cairo, parallel

Bioconductor packages: IRanges, BSgenome, Rsamtools, rtracklayer, GenomicRanges and Biostrings

JavaScript and CSS

jQuery framework - http://jquery.com

Bootstrap - http://getbootstrap.com

DataTables, Table plug-in for jQuery - http://www.datatables.net

jQuery File Upload Plugin - https://github.com/blueimp/jQuery-File-Upload

jQuery throttle - http://benalman.com/projects/jquery-throttle-debounce-plugin/

jQuery Cookie Plugin - https://github.com/carhartl/jquery-cookie

Modernizer JS library - http://modernizr.com

JavaScript Templates - https://github.com/blueimp/JavaScript-Templates

JavaScript Color Picker - http://jscolor.com

md5-js - https://github.com/wbond/md5-js

Font Awesome - http://fortawesome.github.io/Font-Awesome

Google Fonts - https://www.google.com/fonts

jQuery user interface - http://jqueryui.com (documentation)

jquery.tocify.js: jQuery Table of Contents - https://github.com/gfranko/jquery.tocify.js (documentation)

Strapdown https://github.com/arturadib/strapdown (documentation)

Bootswatch themes - http://bootswatch.com (documentation)

google-code-prettify - https://code.google.com/p/google-code-prettify (documentation)

marked - https://github.com/chjj/marked (documentation)

Important conceptual contribution to the project

Liu T, Ortiz J, Taing L, Meyer C, Lee B, Zhang Y, Shin H, Wong S, Ma J, Lei Y, et al. 2011. Cistrome: an integrative platform for transcriptional regulation studies. Genome Biology 12: R83.

Thomas Williams, Colin Kelley and others (2010). Gnuplot 4.4: an interactive plotting program. URL http://www.R-project.org/.

Kent, W.J., Sugnet, C.W., Furey, T.S., Roskin, K.M., Pringle, T.H., Zahler, A.M. and Haussler, a. D. (2002). The Human Genome Browser at UCSC. Genome Research. 12:996-1006.

Kent WJ, Zweig AS, Barber G, Hinrichs AS, Karolchik D. (2010). BigWig and BigBed: enabling browsing of large distributed datasets. Bioinformatics. 1;26(17):2204-7

Nicol, J.W., Helt, G.A., Blanchard, S.G., Raja, A. and Loraine, A.E. (2009). The Integrated Genome Browser: free software for distribution and exploration of genome-scale datasets. Bioinformatics (Oxford, England). 25:2730-1.

Thorvaldsdottir, H., Robinson, J.T. and Mesirov, J.P. (2012). Integrative Genomics Viewer (IGV): high-performance genomics data visualization and exploration. Briefings in bioinformatics. bbs017

Server deployment

Shiny Server - https://github.com/rstudio/shiny-server

ShinyApps - https://github.com/rstudio/shinyapps

Publications containing figures made by SeqPlots

Chen RA, Stempor P, Down TA, Zeiser E, Feuer SK, Ahringer J. Extreme HOT regions are CpG-dense promoters in C. elegans and humans. Genome Res 24(7):1138-1146 Jul 2014